enhancement: moved useQuery uses from components to custom hooks
